Backmerge block renaming fixes/refactors from 6.4 branch into Gutenberg trunk (#56386)
* Apply refactor from 6.4 branch

* Apply fix to lock down rename

* Update support hook to check correct support value

* Rename hook to match feature

* Correct test name

* Assert against a Block that doesn’t support the feature

* Ensure presence of options menu before Rename presence assertion

* Avoid redundant creation of blocks and use Group directly

* Add additional test coverage for options menu on unselected blocks

* Update hook to check for correct support
