+++ packages/block-editor/src/components/block-rename/use-block-rename.js
@@ -1,0 +1,10 @@
+/**
+ * WordPress dependencies
+ */
+import { getBlockSupport } from '@wordpress/blocks';
+
+export default function useBlockRename( name ) {
+	return {
+		canRename: getBlockSupport( name, 'renaming', true ),
+	};
+}
